+++ Sources/World.swift
@@ -361,7 +361,7 @@ public final class World {
             body.derivePositionAndAngle(elapsed)
             
             // Only update edge and normals pre-accumulation if the body has 
-            // components
+            // components - only components really use this information.
             if(body.componentCount > 0) {
                 body.updateEdgesAndNormals()
                 
@@ -370,7 +370,6 @@ public final class World {
             }
             
             body.integrate(elapsed)
-            body.updateEdgesAndNormals()
             
             body.updateAABB(elapsed, forceUpdate: true)